+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_module_functions.go
@@ -11,6 +11,7 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
+	"github.com/kurtosis-tech/container-engine-lib/lib/uuid_generator"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io"
@@ -50,29 +51,17 @@ func (backend *KubernetesKurtosisBackend) CreateModule(
 	image string,
 	enclaveId enclave.EnclaveID,
 	id module.ModuleID,
-	guid module.ModuleGUID,
 	grpcPortNum uint16,
 	envVars map[string]string,
 ) (
 	newModule *module.Module,
 	resultErr error,
 ) {
-	//TODO This validation is the same for Docker and for Kubernetes because we are using kurtBackend
-	//TODO we could move this to a top layer for validations, perhaps
-
-	// Verify no module container with the given GUID already exists in the enclave
-	preexistingModuleFilters := &module.ModuleFilters{
-		GUIDs: map[module.ModuleGUID]bool{
-			guid: true,
-		},
-	}
-	preexistingModules, err := backend.GetModules(ctx, enclaveId, preexistingModuleFilters)
+	guidStr, err := uuid_generator.GenerateUUIDString()
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting preexisting modules with GUID '%v'", guid)
+		return nil, stacktrace.Propagate(err, "An error occurred generating a UUID string for module with ID '%v'", id)
 	}
-	if len(preexistingModules) > 0 {
-		return nil, stacktrace.NewError("Found existing module container(s) in with GUID '%v'; cannot start a new one", guid)
-	}
+	guid := module.ModuleGUID(guidStr)
 
 	enclaveAttributesProvider:= backend.objAttrsProvider.ForEnclave(enclaveId)