+++ container-engine-lib/docs/changelog.md
@@ -8,6 +8,10 @@
     * `GetUserServices`
     * `StopUserServices`
     * `DestroyUserServices`
+    
+### Breaking Changes
+* Removed `ModuleGUID` argument in `KurtosisBackend.CreateModule`
+  * Users will need to remove the argument on each call, the module's GUID will be automatically created in the backend for them
 
 ### Changes
 * Upgraded Kubernetes client SDK from v0.20 to v0.24